Fix accounts sorting
